requirements.txt
click==5.1
ordereddict==1.1
python-dotenv==0.3.0
requests==2.8.1
wheel==0.24.0
